Create find-a-phrase.py
A form of this script was used to look for certain phrases in archival finding aids that were deemed problematic/culturally inappropriate. The example 'passed away' is present in the script, and is a phrase we actively searched for. This phrase is often considered inappropriate when speaking of death and mourning in Jewish culture, and so to ensure that this was not used in records pertaining to the Jewish community an audit was performed.